+++ packages/botbuilder-adapter-twilio-sms/src/index.ts
@@ -7,3 +7,4 @@
  */
 
 export * from './twilio_adapter';
+export * from './botworker';