+++ packages/botbuilder-adapter-web/src/web_adapter.ts
@@ -157,7 +157,9 @@ export class WebAdapter extends BotAdapter {
                     return ws.terminate();
                 }
                 ws.isAlive = false;
-                ws.ping('', false, true);
+                ws.ping('', false, () => {
+                    // noop
+                });
             });
         }, 30000);
     }