+++ packages/botbuilder-adapter-webex/src/webex_adapter.ts
@@ -393,7 +393,7 @@ export class WebexAdapter extends BotAdapter {
                 } else if (activity.channelData && activity.channelData.attachments) {
                     message.attachments = activity.channelData.attachments;
                 }
-                
+
                 // @ts-ignore ignore this webex specific field
                 if (activity.conversation && activity.conversation.parentId) {
                     // @ts-ignore ignore this webex specific field
@@ -401,7 +401,7 @@ export class WebexAdapter extends BotAdapter {
                 } else if (activity.channelData && activity.channelData.parentId) {
                     message.parentId = activity.channelData.parentId;
                 }
-                
+
                 const response = await this._api.messages.create(message);
 
                 responses.push(response);