+++ packages/generator-botkit/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 4.10.0
+
+* Update to Botkit 4.10
+
 # 4.6.3
 
 Fix issue introduced in 4.6.2 that removed required .env file templates